+++ src/extension.ts
@@ -1,3 +1,4 @@
+// Shift6 Formatter by Levent Akdogan (Lakdogan)
 import * as vscode from 'vscode';
 
 export function activate(context: vscode.ExtensionContext) {